requirements.txt
requests==2.25.1
pymongo==3.11.2
dnspython==2.0.0
flask==1.1.2
numpy
pandas
bs4
flask
python-dotenv>=0.8
PyYAML>=5.4